* compiles on qt 4.2 * display line and column numbers in the status bar (thx 4.2 ;)     is it ok if we depend on 4.2? (can I remove the #ifdef QT_VERSION)

+++ qshaderedit/src/parametermodel.cpp
@@ -103,7 +103,7 @@ QVariant ParameterModel::headerData(int section, Qt::Orientation orientation, in
 QModelIndex ParameterModel::buddy(const QModelIndex & index) const
 {
 	if (index.column() == 0) {
-		return createIndex(index.row(), 1, index.internalId());
+		return createIndex(index.row(), 1, (qint32)index.internalId());
 	}
 	return index;
 }
